hello world in vanilla js
