--- HelloWorld.cpp
@@ -1,8 +1,0 @@
-#include <iostream>
-using namespace std;
-
-int main()
-{
-    cout <<"\nHalo Dunia"<< endl;
-    return 0;
-}
