--- HelloWorld.html
@@ -1,16 +1,0 @@
-<!doctype html>
-<html>
-
-  <head>
-  
-    <title> Hello World! </title>
-  
-  </head>
-  
-  <body>
-  
-    <h1> Hello World! </h1>
-  
-  </body>
-
-</html>
