--- Hello_World.py
@@ -1,3 +1,0 @@
-def hello_world():
-    print('Hello, world!')
-hello_world()
