+++ hello-world.py
@@ -1,1 +1,1 @@
-print('Hello, world!')
+print("Hello, world!.......nah,too cheesy;")
