+++ helloWorld.html
@@ -1,8 +1,16 @@
+<!doctype html>
 <html>
-<head>
-<title>Hello World</title>
-</head>
-<body>
-<h1>Hello World!!</h1>
-</body>
+
+  <head>
+  
+    <title> Hello World! </title>
+  
+  </head>
+  
+  <body>
+  
+    <h1> Hello World! </h1>
+  
+  </body>
+
 </html>
