+++ hello_world.py
@@ -1,1 +1,3 @@
-print("Hello World")
+def hello_world():
+    print('Hello, world!')
+hello_world()
