+++ saying_hi_in_js.js
@@ -1,0 +1,1 @@
+console.log("Hello World!");
